+++ src/ui/components/LoginStorybook/Login.css
@@ -46,7 +46,7 @@ body{
     padding: 0.375rem;
     margin-top: 0.625rem;
     box-shadow: 0rem 0.25rem 0.75rem rgba(55, 81, 255, 0.24);
-    border-radius: 05 rem;
+    border-radius: 0.5rem;
 }
 .submit:hover{
 	opacity: 0.8;
