+++ src/ui/components/LoginStorybook/assets/images/logo.png
