+++ src/ui/components/LoginStorybook/components/label/Label.stories.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import  Label  from './Label';
+export default {
+    title: 'ui/components/loginstorybook',
+    component: Label,
+}
+const Template = args => <Label  {...args}/>
+export const LabelLogin= Template.bind({});
+LabelLogin.args = {
+    text:'login'
+
+}
